fundamental/python/requirements.txt
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-mock==3.10.0
SQLAlchemy==2.0.15
